[FIRRTL] Fix Twine usage
When parsing asserts we have a helper class, `ExtractionSummaryCursor`,
to keep track of error messages.  The class holds a `const Twine &`
which created an issue with dangling references.  The following function
was directly creating a dangling reference by passing in `{}` for the
second argument.

```c++
template <typename JsonType>
ExtractionSummaryCursor<JsonType>
makeExtractionSummaryCursor(Location loc, JsonType jsonValue) {
  return ExtractionSummaryCursor<JsonType>{loc, {}, jsonValue};
}
} // namespace
```

By changing the member type to a regular `Twine` the class is
effectively a wrapper around `Twine` and has the same usage
restrictions.