[FIRRTL] Fix use-after-free of ops erased during assert parsing
Fix an issue where the `foldWhenEncodedVerifOp` function would erase
operations while the `LazyLocationListener` of the parser still has a
pointer to them. This causes `endStatement` to try and assign locations
to those operations, which for the erased operations is a use after
free. This adjusts the `foldWhenEncodedVerifOp` function to only remove
the operation but not immediately destroy it. This allows the listener
to do its updates and properly clean up the removed operations once it's
done.

Fixes #2355.