[Handshake] Only buffer merge-like outputs in the 'cycles' buffering strategy (#2356)
Previously, the DFS based approach would naively place a buffer on every site that was determined a cycles. Cycles should only occur at merge-like operations; due to this, buffers would essentially be placed on all of the inputs to a merge-like operation, instead of placing it on the output of the merge-like op.

In this commit, we insert a buffer at the output of a merge-like operation whenever the op is determined to be within a cycle. This greatly reduces the number of buffers placed by the "cycles" strategy.

Additionally, cleaned up some of the buffering code to extract buffer placement routines to separate functions.
A follow up commit will remove all of the buffering passes from StandardToHandshake to be Handshake dialect passes.