[Comb] Don't canonicalize muxs to array-selects when the index is large (#2354)
When checking if we can fold a series of `mux`es into an array+select,
this limits the number of array elements to 512.  This is fixing an
overflow when the bitwidth was 64, causing us to try to store a
`1<<bitwidth` in a `uint64_t`, while also trying to pose a sensible
limit.  The limit was chosen arbitrarily.