[FIRRTL] Avoid inefficient module walks in EmitOMIR
The `EmitOMIR` pass has two places where the entire contents of a
`FModuleOp` are walked just to find a single instance by name.

This PR adds a lookup table for instances by parent module and name,
which is constructed during the initial, unavoidable walk of the IR that
the pass has to perform anyway. The other walks are then removed in
favor of a lookup into this table.

This also restructures the removal of temporary NLAs to have better
scaling by building a set of instance ops with NLA participation and a
set of all NLA symbol names to be removed, and then walking the set of
instance ops and removing annotations referring NLAs in the second set.

Fixes a significant scaling problem in EmitOMIR.