[Handshake] Add buffer on all merge outputs
As far as i interpret the theory, buffers are only needed to break combinational cycles in the style of dataflow circuits that we generate. However, i am starting to suspect that there is something within our flow which requires buffers to be placed on all "logical" cycles as well. Deadlock has been observed in some circuits where control flow is a loop with multiple basic blocks in the loop (b0->b1->b2->b3->b1). This might also be correlated with memory accesses within those blocks, but further investigation is needed.

Nevertheless, experimentally, almost all deadlocks are removed when we add a sequential buffer to break combinational cycles, and for any remaining merge-like operations, add a FIFO buffer on their output, which alleviates the situation described above.