[SchedulingAnalysis] Support scf::IfOp and memref::LoadOp/StoreOp.
This generalizes the auxiliary dependences added for control flow to
support both Affine and SCF IfOps. Similarly, the auxiliary dependence
from the stores to the terminator are generalized to support both
Affine and MemRef StoreOp. With these enhancements, this analysis now
supports the SCF and MemRef dialects, provided the memory dependences
are correctly assigned.