Rewrite to modify field names while printing
With this commit, field names are renamed while their name being printed.
We rename field names in the module scope, not in global scope.
Therefore, it might be possible to emit different field names for the
same type if they are in different modules. It does not cause
compile errors because inputs of verilog module will be implicitly cast.