[HWCleanup] Change the order of simplifications (#2403)
This commit changes the visiting order of simplifications to handle
nested regions properly. Currently `runOnGraphRegion` executes the
following procedure:
```
for op in body:
    recurseToSubegions(op)
    mergeIfPossible(op)
```
This commit modifies the procedure like this:
```
for op in body:
    mergeIfPossible(op)
for op in body:
    recurseToSubregions(op)
```

This effectively handles nested regions. Interestingly, this also
makes `opsToRevisitRegionsIn` unnecessary because we merge
the outer level first.