[FIRRTL] Migrate ExtModule to use ParamDeclAttr
Change FIRRTL's ExtModule to use ParamDeclAttr for parameter storage as
opposed to using a dictionary.  Modify parsing/printing to put
parameters in a carat-delimited area before ports (just like how HW
Dialect does this).

This makes NO modifications to the way ExtModule is represented.
Parameter values are still stored on the ExtModule and not on the
instantiation.  This will be changed in a later commit.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>