[ExportVerilog] Legalize field names of struct types (#2359)
This commit adds `FieldNameResolver` class to legalize field names in
struct types. Field names may contain verilog reserved keywrods so it's
necessary to rename them before the emission. 

With this commit, field names are renamed while their name being printed.
We rename field names in the module scope, not in global scope.
Therefore, it might be possible to emit different field names for the
same type if they are in different modules.