[FIRRTL] Add moduleNameAttr to FModuleLike
Add a utility method to FModuleLike that returns the _attribute_ of the
module/extmodule name.  Previously, there was only a "moduleName" member
function which returned a StringRef.  Adding a way to get the actual
attribute complements patterns where you need an attribute, e.g., as a
key to a dictionary.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>