[FIRRTL] Add ability to get the type ultimately targeted by a fieldID
We have a function to return the type of a subfield using a fieldID.
The problem with this function is that it only strips a single layer at
a time from aggregate type, and it doesn't return the index of the
target element or its field ID.

This changes the functionality to return both the subtype, and the next
fieldID.  This makes it easier to iteratively follow subtypes all the
way to the final type.

This also adds a function to return the type targeted by the fieldID and
skip manually walking the data structures.