[FIRRTL] Move AnnoTarget out of LowerAnnotations, fix bug
This moves the `AnnoTarget` class out of the `LowerAnnotations` pass.
This is so that it can be used in the `Dedup` pass.  The point of this
class is to track what an annotation is targeting, namely a certain op
or a port on an op.

The class was changed to use sub-types for representing ports with
`PortAnnoTarget`s or ops with `OpAnnoTarget`s.  The base class
`AnnoTarget` can be used with `TypeSwitch` to downcast to the derived
type.

The `FieldID` was moved out of the `AnnoTarget` class, since the
`Annotation` class already captures this information. `LowerAnnotations`
required some larger changes to work with this change.

This also fixes a bug in `LowerAnnotations` where it would only resolve
1 field deep of nested aggregate types.