[FIRRTL] Handle NonLocalAnchor on Bundle fields (#2513)
This commit adds the missing symbols for fields of a Bundle.
We must add a symbol on the operation which has a `Nonlocal` anchor associated
 with any of its fields.
Parser was not adding symbol when the `circt.nonlocal` is present on a subfield
 of a Bundle type.

The `LowerTypes` must then fixup the `NonLocalAnchor` after lowering, because
 the lowered bundle fields will get new symbols. 
This requires the NLA to be updated with the new lowered symbol.

`LowerTypes` should record all the bundle fields which have a reference to the NLA, and the corresponding 
symbol for the subfield. The NLA is then updated at the end of the pass.