[StaticLogic] Add start time attribute to PipelineStageOp. (#2556)
This is an essential result of scheduling, and should be captured in
this IR. It is a property of the pipeline stage, not the operations or
operator types they might map to. This information, with the pipeline
II, is needed by lowering passes to generate the appropriate
cycle-level timings. To start, it will be used to determine which
stages need to be considered in the pipeline prologue and epilogue.