[FIRRTL] Add a data structure for recording InnerSyms. NFC. (#2573)
This commit adds a data structure to record and lookup an Inner symbol and the
corresponding operation. 
An inner sym constitutes of the module name and the symbol name.

The data structure is implemented using SmallVector instead of HashMap. 
It can be used when the list is populated first, then sorted and then only used for lookup.
The assumption is that the list must be sorted, before any lookup.
This does not handle duplicate entries explicitly.

The idea is based on an observation that Dense arrays can be efficient lookup
structures. Especially when we have insert-phase and lookup-phase based code.

This is currently being used for NLA verification (#2461).
This commit is part of PR #2461.

TODO:
Make the list more general.
Some runtime comparisons with DenseMaps must be done, to investigate
any performance improvement.