[FIRRTL] Fix printf-encoded lint issues (#2547)
Fix a bug when parsing certain types of printf-encoded
asserts ("assert:", "assume:", "cover:" and "assertNotX:") to not add an
unnecessary format string argument.  Instead, only add argumnets after
the first argument if any exist.  The latter case is extremely rare (to
use this from the Chisel APIs you would have to hand-craft a printf
operation and not use the built-in helper utilities).

Change Chisel asserts ("Assertion failed:") to always parse as immediate
assertions.  These assertions are expected to have format string
arguments and this avoids a catch-22 with VCS lint warnings.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>