[FIRRTL] Prefix everything in the ModuleInliner
Change the ModuleInliner to put the inlined prefix on any operation
which has a "name" as opposed to just on a list of operations known to
FIRRTL to have a name.  This helps future-proof the pass as well as
keeping naming consistent if new operations or totally unknown
operations are seen.

h/t @darthscsi

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>