[FIRRTL][InstanceGraph] Make instance graph mutable (#2550)
The instance graph was storing graph nodes in an array and handing out
pointers to the elements, which meant we could not properly delete nodes
from the graph without invalidating a bunch of node pointers.  To make
the instance graph more easily mutable, this changes the underlying
data structure of the graph to have stable pointers.

This change makes the instance graph follow the same model as
operations:  InstanceGraph holds a intrusive list of InstanceGraphNodes,
and InstanceGraphNodes hold an intrusive list of InstanceRecords.  The
uses are kept track of using a doubly linked list which is intrusive in
the InstanceRecord.  One benefit of using a representation similar to
MLIR operations is that we can closely match the API, which has slightly
less mental overhead when mutating the instance graph.