[SCFToCalyx] Finish initial pipeline lowering implementation. (#2558)
This includes three major additions:

* The BuildPipelineRegs pattern, which establishes pipeline registers
for each stage's results
* The BuildPipelineGroups pattern, which creates groups for each stage
to assign the registers
* Modifications to BuildControl to do pipeline-specific things that
deviate from how scf.while is handled

BuildPipelineRegs occurs at the same phase as BuildBBRegs, before
general ops are lowered in BuildOpGroups. This ensures all arith ops
are correctly lowered into register transfers.

After that, BuildPipelineGroups goes through the generated groups, and
acts differently depending on whether BuildOpGroups created a
combinational or sequential group. If a combinational group was
created, it is turned into a sequential group that writes to the
pipeline register. If a sequential group was created, it is updated to
write to the pipeline register, and the auto-generated register is
removed.

Finally, in BuildControl, some changes are made to achieve a schedule
specific to pipelines. The wrapper SeqOp is removed around pipeline
blocks, and they are wrapped in a ParOp. Additionally, a prologue and
epilogue are added to prime and drain the pipeline.

Together, these changes allow a program like the dot product test to
be lowered through Calyx to Verilog successfully.