[FIRRTL] FIx memory port NLAs + verifier
When we attach non-local annotations to memory ports, since we do not
assign port symbols to the memory op, the leaf element of the anchor
should target a symbol on the memory.

There was a mistake in the LowerAnnotations pass where an `AnnoTarget`
should have been passed by reference, which resulted in memory port
annotation being attached directly to the memory instead.

The NLA verifier checks to see that there is an annotation using the NLA
attached to an operation.  When the target operation is a memory port,
the NLA verifier also has to check the port annotations.  This same
logic would apply to InstanceOps, but we have decided that InstanceOp
annotations should be moved to the target module.

The old/default annotation parsing framework is still broken: it creates
the NLA for the port annotation but does not attach a symbol to the
memory operation.