[MSFT] Update partition pass global refs for inner nodes (#2601)
The previous algorithm to update global refs after partitioning would
only work if the leaf entities were both tagged for placement with
global refs, and targetted to the design partition. In general, we
need to support targetting arbitrary points in the hierarchy to the
design partition, and fix up any global refs pointing to these nodes
and anything below them.

Rather than assuming we are starting from the leaf of the path in
bubbleUpGlobalRefs, we find the correct node to start from. Similarly,
in pushDownGlobalRefs, we cannot assume we are starting from the leaf,
and must find the correct point to insert the nodes.