[ExportVerilog] Change unnamed from "_T" to "_GEN"
Change the name used when ExportVerilog generates a name for an unnamed
things from "_T" to "_GEN".  "_T" is currently used by Chisel for
temporaries it generates and "_GEN" is currently used by the Scala
FIRRTL Compiler for temporaries it generates.  This will benefit Chisel
users compiling with CIRCT as this will continue the clear delineation
between "Chisel generated a temporary" and "the FIRRTL compiler
generated a temporary".

Fixes #2571.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>