[HW] Allow OutputOp outside of modules
Remove the constraint on `OutputOp` that its parent has to be a
`HWModuleOp` specifically. This makes the operation reusable in contexts
of other dialects which introduce their own module-like operations with
similar semantics to the HW modules. An example is an operation
introducing a nested region into the module for some reason, where using
the already existing `OutputOp` would be highly desirable.

As a result, the operation containing the `OutputOp` must merely provide
a `FunctionType` attribute named `type`, pretty much like any other
builtin-style function in MLIR. In the long term, we might want to add
an interface for ops that can act as parent to an `OutputOp`, which we
can use to query the required output type signature.