[FIRRTL] Donot add NLA for OMIR, if already exists. (#2633)
This commit checks for existing `circt.nonlocal` annotation and reuses the 
`NonlocalAnnotation` op for SRAM OMIR. This case occurs, when dedup results in
multiple instance paths to the SRAM.
If there are multiple instance paths, that implies the annotation must be nonlocal,
and the NLA must already exist. 
This commit fixes a crash with `-dedup=1`, which results in the multiple instance 
paths.