[StaticLogic] Add optional trip count attribute.
This is a useful piece of high-level information that is intrinsic to
the pipeline. This could be used by the Calyx lowering to easily
generate a @bound annotation when possible. It could also be useful to
generalize the Affine lowering to nested loops, by iteratively
applying the current analysis and transformation.