[SCFToCalyx] Support optional trip count attribute on pipelines.
This adds a getBound method to WhileOpInterface. For PipelineWhileOp,
the optional trip count attribute is returned. If a bound is
specified, a bound attribute is added to the WhileOp. This is already
supported by the Calyx emitter. The bound attribute is required on
WhileOps for the native Calyx compiler to infer static timing.