[FIRRTL][Dedup] Fix creation of symbols with empty name
When creating an `inner_sym` for ops, Dedup was using the "name"
attribute of the op as a base.  For many ops, it is possible that the
attribute is an empty string, which should not be used to create a
symbol.  This change uses `inner_sym` as the symbol base when the name
is empty.  This does not change the behaviour for ports: they must have
a non-empty name.