[FIRRTL][Dedup] Fix issue with NLAs on ExtModule ports (#2714)
ExtModule deduping has an optimized version of NLA fixup where it
assumes that the ExtModule is the last element of the NLA, since there
can be no wires or other instances inside the body.  This function was
assuming that the NLA was a module reference and would accidentally take
port references and change them into a module reference.

Properly managing port references in NLAs involves tracking the merging
or renaming of port symbols between the two modules.  Since this is a
bit tricky, the best thing to do is use the same NLA fixup code used on
regular modules.

This change also makes some port annotations stronger by using different
names between the two deduped modules: no code-changes were required for
this.