[FIRTL][Annotations] Fix issue when module name matches inner name (#2717)
The annotation parser was mistakenly getting confused when a wire (or
instance) inside a module had the same name as the module itself, and
would create a reference to the module instead of an innerRef to the
wire.

This was not that it was mistakenly targeting the Module per-say, but it
was a bug in the NLA creation logic that used `targetName == moduleName`
as a sentinal value to insert a module reference.  Instead we use
`targetName == ""` as a sentinal value.