[HW] Add port modification for module-like ops
Add the `modifyPorts` function to `HWModuleOp`, `HWModuleExternOp`, and
`HWModuleGeneratedOp`, plus convenience wrappers `insertPorts` and
`erasePorts`. This function performs insertion and removal of input and
output ports in one go, and updates the type and name attributes
accordingly. Very handy when a pass actually needs to modify the
interface of a module. Block arguments are left untouched at the moment,
as the expectation is that the user will have to do some special
handling of those anyway while they are updating the module body. Maybe
in the future these functions could support an opt-in "please take care
of the block args for me" option.