[FIRRTL][LowerTypes] Fix std::lower_bound usage
`std::lower_bound` does not return `end()` when the searched for element
is not found: it can return any element `<=` the searched for element.
For this reason it is necessary to check for equivalence after checking
that an element was found.

This was causing a crash in several SiFive cores although I have not
been able to reduce it to a simple test case yet.  The cause of the
crash appears to be that the `it->op` is an invalid op address. It is
unclear to me how an invalid op made it into the `opSymNames` list in
the first place, but maybe someone more familiar with the code can
figure it out.