[LowerToHW] Don't wrap RANDOM define in brackets
Change defines for register randomization to define "RANDOM" as
"$random" and not "{$random}".  The Scala FIRRTL Compiler (SFC) emits
the former and certain tools (Verilator) want all defines to be exactly
the same.  Chisel users that are mixing Chisel-produced Verilog with
handwritten Verilog can run into tool issues when switching from the SFC
to CIRCT if they have handwritten Verilog that defines "RANDOM" as
"$random" which will conflict with the previous CIRCT definition of
"{$random}".

The old brackets are moved from the definition to each use.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>