[FIRRTL] Verified types for RegOp, NodeOp and RegResetOp
This patch introduces a new type, `NonAnalogPassiveType`, to represent
the values that can be assigned to nodes and registers. Section 4.1.4
bans `AnalogType` from being directly used, although analog types in bundles
and vectors are still permitted and asserted by tests included in CIRCT.
Section 6.9 requires the type of Muxes to be passive, so `PassiveType`
is still required.

Additionally, the type of the reset value of a register must match
the type of the register, according to the language of the spec.
While the specification is vague, the FIRRTL compiler expects
the types to be equivalent structurally: bundles and vectors must
match, although both integer truncation and extension is allowed,
verified by existing CIRCT tests.

The spec could be changed to recursively ban analog, changing the
condition of `NonAnalogPassiveType` to `containsAnalog`. Additionally,
truncation in register initialisation could be banned to define
behaviour corresponding to `ConnectOp`.